+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -239,7 +239,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KDX3CMG794J13G8RXE5GJRD0"
+        "01KDYVT019Q59ZKVNRDF71KDGN"
     )
     r"""The unique identifier of the data source"""