+++ packages/orq-rc/src/orq_ai_sdk/models/generateconversationnameop.py
@@ -152,7 +152,7 @@ class GenerateConversationNameResponseBody(BaseModel):
     r"""Unix timestamp (in milliseconds) when the conversation was last updated"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kdx3cm33cgsyfmr953gsx8df"
+        "conv_01kdyvszm0jws32c2xbjrmamf2"
     )
     r"""Unique ULID (Universally Unique Lexicographically Sortable Identifier) for the conversation, prefixed with \"conv_\" """