+++ packages/orq-rc/src/orq_ai_sdk/models/updatecontactop.py
@@ -181,7 +181,7 @@ class UpdateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-01-01T15:38:46.209Z")
+    updated: Optional[datetime] = parse_datetime("2026-01-02T08:04:42.035Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")