+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KDX3CMGAT9KY0A63V2YSJZ0P"
+        "01KDYVT01C3YWH5F1R27PVYFH0"
     )
     r"""The unique identifier of the data source"""