+++ packages/orq-rc/RELEASES.md
@@ -4509,3 +4509,13 @@ Based on:
 - [python v4.2.0-rc.6] packages/orq-rc
 ### Releases
 - [PyPI v4.2.0-rc.6] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.6 - packages/orq-rc
+
+## 2026-01-02 08:07:27
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.680.11 (2.788.15) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.2.0-rc.7] packages/orq-rc
+### Releases
+- [PyPI v4.2.0-rc.7] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.7 - packages/orq-rc